custom-elements/creating-and-passing-registries/existing-registry-test-001.html

<!DOCTYPE html>
<html>
<head>
<title>A new document instance must be associated with an existing registry if createDocument or createHTMLDocument methods are invoked</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="When DOMImplementation's createDocument method is invoked with namespace set to HTML Namespace or when the createHTMLDocument method is invoked, use the registry of the associated document to the new instance">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#creating-and-passing-registries">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    var name = 'x-frame';

    var GeneratedConstructor = doc.register(name);
    var doc2 = doc.implementation.createHTMLDocument('Document 2');
    //FIXME Should the type of the exception be specified here?
    assert_throws(null, function () {
        doc2.register(name);
    }, 'Exception should be thrown for custom element, which is already registered in shared registry');

    var xframe = doc2.createElement(name);
    assert_true(xframe instanceof GeneratedConstructor, 'Wrong custom element constructor');

}, 'Document created by createHTMLDocument should share an existing registry');


test(function() {
    var doc = newHTMLDocument();
    var name = 'x-frame-1';

    var GeneratedConstructor = doc.register(name);
    var doc2 = doc.implementation.createDocument(HTML_NAMESPACE, 'html', null);
    //FIXME Should the type of the exception be specified here?
    assert_throws(null, function () {
        doc2.register(name);
    }, 'Exception should be thrown for custom element, which is already registered in shared registry');

    var xframe = doc2.createElement(name);
    assert_true(xframe instanceof GeneratedConstructor, 'Wrong custom element constructor');
}, 'Document created by createDocument with HTML namespace should share an existing registry');

</script>
</body>
</html>
